Major work on katepart for kde4 begins.
This is mostly my work from the 3.x series that I never committed due to time
and QA reasons.  In summary:
- Refactored some of the smart cursor/range stuff.  Made it a bit more
  generic in places.  Moved classes to their own files.
- Make KateLineRange more of a class than a struct.  Reasons for this
  will become clearer throughout the bidi stuff.  This is going to be
  the major communication method between the view and the renderer.
- Some QT4 porting (setMicroFocusHint)
- Hooks to allow highlighting to respond to mouse position / movement :)
- Arbitrary highlighting broken at the moment.  Don't worry, plenty of
  time

It's a bit rough around the edges as discussed with Cullman... my gdb is
out of whack at the moment.

svn path=/branches/work/kde4/kdelibs/; revision=418372

+++ kate/part/katehighlight.cpp
@@ -1455,7 +1455,7 @@ void KateHighlighting::doHighlight ( KateTextLine *prevLine,
       if (item->region2)
       {
         // kdDebug(13010)<<QString("Region mark 2 detected: %1").arg(item->region2)<<endl;
-        if ( !foldingList->isEmpty() && ((item->region2 < 0) && (*foldingList)[foldingList->size()-2] == -item->region2 ) )
+        if ( !foldingList->isEmpty() && ((item->region2 < 0) && (int)(*foldingList)[foldingList->size()-2] == -item->region2 ) )
         {
           foldingList->resize (foldingList->size()-2);
         }
